code cleanup
Reindented HelloApplication with 4 spaces

Removed unused methods from UserRepository

Renamed a method in Hello. The old name did not convey the purpose of
the method well.

Changed id parameter used for User object construction from 1 to 0, in
order not to suggest it is going to be id 1, whereas in reality the id
is later autogenerated anyway, in persistence layer. Therefore 0 seems
less misleading.

Removed unnecessary REST mapping from Hello class.